Create change
Patch-set: 1
Change-id: Iff69e4a936f76f19c45dbcc515d17941d0c12096
Subject: Show remaining SMS letters on the "Send" button, from the first letter on. This removes the TextView field that used to take the numbers. Also, the remaining letters are displayed as soon as text is typed. In case of an MMS the "MMS" label is shown.
Branch: refs/heads/froyo
Commit: 96d915fbefe9340e6510cb019596bbeb079d027a