Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Works without issue, but I think from an aesthetics point of view it'd be better to just have the stock TextView always be visible. What was the reason behind putting all the info in the button?

Also, please reformat your commit message; the first line should be <= 60 characters and be a quick summary of the commit. Then skip a line and explain the change in detail.

Patch-set: 1