Update patch set 1
Patch Set 1:

Thanks for the feedback. I looked into other possibilities again today. And i came to the conclusion that i'll stay with this patch. Sorry, but adding a config option for this or supporting both display routines is a bit of overkill in this case.
This solution gets rid of the extra ensureCorrectButtonHeight method and makes the display of remaining letters also work in landscape mode with fullscreen keyboard enabled.
So again, I would be greatful to see this merged.

Patch-set: 1