Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

(1 inline comment)

It does look good to me. We should test it with a few operators to make sure how it behaves in different carriers and locations.

Also, we need to take into account the fact that a user may have the wrong timezone configured. It happens all the time during daylight savings.

To be honest, what I'd really like to be done is storing _both_ timestamps, and change the app to allow the user to choose which one to present and sort by: Received Date or Sent Date (and default to "Received", the current behavior). But since that implies changing the database structure, I'll settle with an option, (again Received/Sent Date), that influences which date to store (System.currentTimeMillis() or sms.getTimestampMillis())

Patch-set: 1