Update patch set 2
Patch Set 2:

Yeah, I'd like the 2 stamps, too, but we can add that later. 

Merge this one, it'll allow us to get some feedback on the reliability of origin timestamps until we get both stored. That way we can make a better informed decision as to which should be the default.

Patch-set: 2