Update patch set 1
Patch Set 1:

I can play Angry Birds on my Legend so I guess a simple subtraction shouldn't turn out to be that big of a problem. ;)


Seriously, if you don't decide to rewrite the patch I'll do that tonight (or try to) and see if it slows down anything. My biggest thread is only about 100 txts though.

Patch-set: 1