Update patch set 1
Patch Set 1:

When the "Use original timestamp" option was added (uses the sent time instead of received time on received messages), there was a hard coded timezone offset. However, for most people the sent time provided by the operator is already timezone offset. This resulted in a wrong timestamp. This change makes the timezone offset for the sent timestamp optional. I also changed a few strings to make it more clear what the setting does.

Patch-set: 1