Update patch set 4
Patch Set 4:

Essentially to save a few clicks :P

Well it was in stock Samsung roms and i found myself using that a lot. But now when i think of it, the stock way isnt that bad either.

Yea general consensus seems that this doesnt really have a good use case, so "meh" im abandoning it :D

Patch-set: 4