Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Currently there is still logic which hides the counter if the text does not span more than two lines...if currentTextLines <= 2 the counter is hidden.  While we certainly could address this and display the counter at all times, it would increase the height of the text input at all times when enabled, resulting in a less than desirable user experience/interface IMO.  Another place that it is currently hidden is when the message is a MMS message versus a SMS.  Again, we could address as well, but neither of these were addressed in CM7, just let me know how you would like me to proceed.

Patch-set: 1