Add option to display contact pictures in Mms notifications.
According to Google's Android Design guidelines for notifications
(http://developer.android.com/design/patterns/notifications.html):

"For notifications of items sent by another user (such as a message or status update), include that
person's image. Remember to include the app icon as a secondary icon in the notification, so that
the user can still identify which app posted it."

This is not adhered to in the Mms.apk. If a contact with a picture sends an SMS/MMS a picture of
that contact should be displayed. If mulitple conctats send an SMS/MMS, the standard notification
for multiple messages is displayed.

This option can be enabled/disabled via Mms > Settings > Notifications > Contact pictures

Update: removed styles.xml as it was not changed.

Change-Id: I580693fe6ceaae560e4909e7760934c8bbd06f32