Update patch set 1
Patch Set 1: No score

We found that too many users were changing this setting, not knowing the implications. If your carrier does not support un-split SMS, the recipient of the message receives garbage data, while the sender is unaware of the problem. Since the carrier either supports split sms or doesn't, it should be set once in mms_config.xml by the device maintainer.

Patch-set: 1