Update patch set 1
Patch Set 1:

Justin: So the idea is to default it to enabled and then disable it using an overlay for devices on carriers known to not support it?

Was it the mms_config.xml that could have overlays on the MMC and MNC level or am I thinking about another config?

Patch-set: 1