Mms: Fix wrong message selection in context menu
Sometimes when the user open a message's context menu for the
first time, the action will be performed on the incorrect message
(i.e. deleting the wrong message). This is because the menu click
listener relies on the fact that that the cursor points to the
current message, which is not always true because bindView can
be called on other messages causing the cursor to move.

Change-Id: I4e8fdd2956987fc3f12c80def46c107a4089d0ae
Fix: Cache the position of the message and move the cursor back.