Added "Support for emojis" on the MMS app
Support for the Unicode subset dedicated for emojis, as used on the iPhone, or even the Gmail website.

The full list with Unicode codes is available here: http://barrow.io/posts/iphone-emoji/.
The pictures used here are not bound to any licenses, even GitHub is using them (https://github.com/blog/816-emoji).

The size of the icons seems OK to me, but I could only test on an HDPI device (Galaxy S II i9100). MDPI icons are also included, but I couldn't really test it.

A new checkbox preference has been added to be able to toggle the feature.
The feature will be disabled if the "Strip Unicode" feature is enabled (obviously).

PS1: Initial commit
PS2: pngcrush on all the pictures + XHDPI emojis

Change-Id: I8a457498fb655ea3f1bb273d7756f7c224b81466