Added support for Unicode 6.1 emojis
The previous patch used the Softbank private unicode code page to encode and decode emojis.
This was the technique used by iOS before version 5 and all the Japanese carriers.

Google and Apple have proposed a new code page which has been integrated into Unicode 6.0, then updated in 6.1
The full specification can be found here: http://www.unicode.org/~scherer/emoji4unicode/snapshot/full.html

NOTE: If we were to redesign all the icons, we would also need to support the entire code page, which is way bigger than the one Sotfbank used :D

This patch changes the way emojis are encoded, and adds a new decoding table but keep both for compatibility.

Surrogate pairs are used, because of the way Java handles UTF-16 chars - these cannot be included in the XML resource file.
Softbank Unicode code points have also been moved in the EmojiParser.java so everything is in the same place.

Also changed the label to reflect the UI guidelines in the Wiki.

TLDR: It's now standard, and some emojis even show up in Gmail or in Chromium

PS1: Initial patch
PS2: Fixed typo in the Softbank codes

Change-Id: I9eb5081eb7514cb2b89ed3b91cc3054347d6ac90