Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: This patch add an option to disable the transformation from SMS to MMS when the lenght of the message is > 3 sms. It is also possible to still have this transformation but to change the number of text needed before the sms is transformed to mms
Commit: 8fe4fd98370ea284a53451d4ecf444b80200b937