Create patch set 3
Uploaded patch set 3.

Patch-set: 3
Subject: Add an option to disable the transformation from SMS to MMS when the lenght of the message is > 3 sms. It is also possible to still have this transformation but to change the number of text needed before the sms is transformed to mms
Commit: e05c2f35af2bc17689f395c9e09866746640d8c0