Update patch set 1
Patch Set 1:

Danny.

On Aurora for ICS I tried to add this strip to that table, however it didn't work.

After DZO add it this way it started working.

Dzo wrote on his commit: "Looks like the character encoding thinks ç is encodable but Ç is not. It's actually the other way round. So add a check in the unicode stripping function."

Patch-set: 1