Update patch set 2
Patch Set 2:

so wait.. the commit that this reverted will be in jb but doesnt make mms.apk slow? what about the lockscreen thing? ( as a reminder: if lockscreen is set to off with the screen off and u leave message up it still acts as if u r looking at and doesnt notify u )

Patch-set: 2