Mms: fixes battery drainage while receiving Mms with data connection off
When the user receives a pre-Mms (message with the subject and URI), the
TransactionService acquires a partial wake lock, and tries to use the data
connection (in 99.99(9)% of the cases) to donwload the rest of the Mms from
that URI.

But if no data connection is available, the TransactionService requests an
automatic connection to a data connection and sits while waiting for an Intent
saying that a data connection is available so it can download the mms.

Normally, as the system won't connect automatically the data connection
(because of the many bad implications - roaming in a foreign country,
limited data plans, etc.), the TransactionService keeps the wake lock active,
and it does not even warn the user that an Mms is waiting for a data connection.

In the point of view of the user, he just starts to see that his phone starts
draining the battery pretty fast, does not go to deep sleep  and the user
does not know why because he does not receive any notification at all!
He can only do a reboot, or curse that CM9 is pretty battery intensive..

With the data off this patch "catches" the data-connection-not-available
intent, and proceeds to cancel all pending mms downloads. It then notifies the
user, showing the Mms's (without the images, sounds, etc.), and the user can
download the rest manually when a data connection is available, if he wants.
And, as a bonus, it releases the wake lock and lets the phone sleep again.

With data on, the procedure is unchanged: It automatically downloads the mms
contents and notifies the user when everything is downloaded..

So, the changes for what would be normally expected is:
- When the user switches the data connection on, he must download the mms
  manually, as it won't start automatically anymore. But he will have battery!
- When the user sends an mms with the data connection off, it will say that the
  message will be sended when a connection is available. But that will never
  happen, so only send Mms with the data connection on. But again, he will have
  enough battery charge to continue with his life!

Patchset2: Cleaned trailing whitespaces

Change-Id: I319e28c04aadd1f0e309fa5902a05646e77cf60c