Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The entire point of the wakelock is to download all of the messages when data connectivity is reestablished. Yes it drains the battery a bit, but from what I personally see not that much. If a user receives 10 MMS messages (for whatever reason) they will have to download all of them manually.

Also check your code for trailing spaces and tabs. There are a lot in your code and should be removed.

Patch-set: 1