Update patch set 1
Patch Set 1:

The problem is that the wakelock will never be released until the user starts a data connection. So, it will drain the battery as much as the time it takes for the user to switch data connection. For non-tech users, if they don't know that an mms has arrived (no notification was given), the battery will drain ad-eternum. It was what has been happening to me until I realised this, because I never ever switch data connection on. The same happens with non-tech users like my sister and my girlfriend which are being cm9 beta-testers for 1 month now, and complain about battery drainage (2G/no-data-connection)

If a user receives 10 mms and the data connection is on, all will be downloaded. If not, I think he would prefer to have charge in his battery to download later. :)

About tabs, I have Eclipse to change tabs to spaces, so I don't know how to solve it..

Patch-set: 1