Update patch set 4
Patch Set 4:

Attention: The only way to see this working is to compare the before and the after the patch.

Before the patch, if you receive an mms with the data connection off, the phone will be awake (although with the screen off) until you switch on the data connection. Just send an mms from another phone to yourself (data connection off), wait some time and go to the battery statistics (Settings->Battery->(Press the top graphic section) and see that the Awake blue bar is on for quite some time since you received the mms.

With the patch, you will receive a confirmation of reception and the phone does deep sleeps again..

I have not seen any side effects so far, besides battery improvement after receiving mms with data off..

Patch-set: 4