Update patch set 1
Patch Set 1:

Yes we could do that, but it means hours of work, and I'm not a designer. Plus, the glyphs located in AndroidEmoji.ttf won't look good if they are bigger, and even if they are colored I think.

Finally, as I said in the message, this is simply a port of the patch I made for CM9 to CM10 so the features are at the same level.

I'll be happy to discuss about the evolution of the feature, I'll keep digging about the emojis Google included in JB.

Patch-set: 1