Update patch set 1
Patch Set 1:

Yeah, ok. I agree that we can look at that later. This change would be needed if we did that anyway I think.

Just as a demo of how they look at 33px x 33px (the size Google uses for the ones included in Mms.apk).

https://www.dropbox.com/s/7e4njo4cy62qbbw/android_emoji_demo.png

As they are font glyphs, they are vector "images" so they scale nicely to whatever size.

Patch-set: 1