Update patch set 9
Patch Set 9: No score

David, this workaround is not working. I see that the layout is fit with the size of the screen and exceeds the dialog size. I'm reviewing the layouts trying to figure out what is wrong, but for the moment, i have no result. Maybe, something is wrong in my sources that is causing the incorrect alignment. If i find what is wrong i tell you.

Patch-set: 9