Update patch set 1
Patch Set 1:

This is a revert of a change made in 21112. The vibrate on receipt of Mms is broken since that change merge, and only vibrates when the device is set to vibrate mode. It used to work as it was before this change. It seems that change 21112 introduced the bug it claimed to resolve, reverting this restores working vibrate functionality. 

Testing scenario: receive an Mms with the 'vibrate always' mode enabled in Mms settings while your device volume is up (ie not in silent or vibrate mode). With change 21112 the device will not vibrate even though it sends the intent for it. With that change reverted, Mms notification vibrates as expected.

Patch-set: 1