Update patch set 1
Patch Set 1:

Utkanos, Its not that it is broken, its that it does not pay attention to the System Sound setting of "Vibrate when Ringing"  If you use the Notification Widget, and Set it to Ring+Vibe, then the patch works as intended.  I have already contacted Jorge Ruesga about the issue, and I believe he is looking into it.

Patch-set: 1