Update patch set 1
Patch Set 1:

Ah I see, I'd rather custom notification settings in the app not be affected by the ringer settings personally. Sort of redefines what the "always" in "vibrate alway"s means in the mms settings and changes how it's worked for a while. Perhaps it's possible to at least revert this until it can be addressed another way?

Patch-set: 1