Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

The patch is working for me in my devices, even changed the values ??of System/Sound/Vibrate when Ringing. System setting should not affect to this patch. Maybe it works when sound button of notification bar is pushed for first time.

Anyway, i prefer revert this patch because it's causing some kind of problem and don't fit the user's issue for which it was created.
I will revert also patches for Email and Calendar.

Sorry, for upload wrong patches.

Patch-set: 1