Mms: Add Template support to QuickMessage pop-up
This commit adds native Messaging app templates support to the
QuickMessage pop-up.

Clicking on the icon to the left of the reply editor will display
a dialog with the  available templates allowing the user to select
one.  It will then append the selected text to the reply message.

Screenshot:
  https://dl.dropbox.com/u/50993944/Screenshot_2012-09-03-15-43-42.png

Patchset 2  - Remove holo.dark template drawables

Change-Id: I669bf65e60c66d875276a711f76b602f4c3dd395