Update patch set 1
Patch Set 1:

Thanks for review Daniel. Adding msgId to the file name is not required as getUniqueDestination() already ensures that the file name is unique (by adding sequence number at the end of the file). I think foreign characters in file names should work fine as the filesystem uses UTF-8. Anyway I agree it's a good practise not tu use these in file names for portability. The problematic characters are "/'? etc. Do you think it's necessary to add this information in the string resource comment?

Patch-set: 1