Update patch set 4
Patch Set 4:

@Gustavo, I am just leveraging the existing option menu functionality and look and feel of the messaging app.  Using the same UI constructs in the pop-up will be familiar to the user.  As for against design guidelines, not sure that it is.  

My #1 design principle is to keep the pop-up UI as simple as possible and uncluttered.  I am not open to adding more buttons/rows to it.  We have limited space as it is already (especially on lower DPI devices) and one of the main reasons why I did this in the first place is that I cannot stand the UI of GoSms, handsent etc - way too cluttered.

Patch-set: 4