Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

You are right, the quick message shouldn't be bloated on the UI front.

I personally, don't think this is a good idea. We are basically doubling up functions in the messaging app. The quick message setting should be just that, a quick way to respond. If you want to add all these extra bits, then just click name and use the actual app. This is going to over complicate for no reason.

Without there being a dark "theme" in the actual app layout, having one in just the quick message is a little off IMO.

Patch-set: 4