Update patch set 4
Patch Set 4:

I see the ability to insert smiley quickly & easy to use as part of the core functionality to Messaging & they should be included in QuickMessaging. By extent this also included emoji.

I would also like the see the option to be able to switch from Holo.Light to Holo.Dark in the main Messaging app if we're going to be providing options to switch between a light theme & a dark theme to QuickMessage.

Patch-set: 4