Update patch set 5
Patch Set 5:

I don't really see an issue with that string. That is how it was worded in CM9. But, I see where you are coming from and don't really have an issue either way. Is there anyone in particular who would need to make the call on it?

Patch-set: 5