Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

IMO this is not the way we should handle.  This is setting a precedent that we should just start adding preferences for different drawables.  A theme would be the proper way to overwrite the drawables if the user wished.

Patch-set: 1