Update patch set 2
Patch Set 2:

Re: Template insertion, I'm on the fence.  If I understand the intention of templates, they would be used to simply and quickly respond with a full pre-typed message.  Emoji and Smileys differ in that they are 1-3 characters that the user may insert in final editing of a text for emphasis.  

At the same time, it makes sense to keep the operation consistent and there is no loss of functionality if we insert a template at the cursor position in the same way.

I'll work on making the template insertion consistent :-)

Patch-set: 2