Update patch set 5
Patch Set 5:

I think the comments should read "Add SMS to calendar reminder" instead of "Add SMS to calendar remind".

Also, please add a blank line between the title and the description in the commit message so that it's easier to read. I figured you might as well change it if you change the comments.

Patch-set: 5