Update patch set 5
Patch Set 5:

Just tested this on my Samsung Galaxy S and the system popped up a request to use Calendar or Jorte (3rd party calendar app).  Selecting Jorte did not properly fill the entry.  Do we possibly need to use Intent.ACTION_INSERT instead of Intent.ACTION_EDIT?  When I substituted ACTION_INSERT I was directed right to the calendar entry screen.  http://developer.android.com/guide/topics/providers/calendar-provider.html#add-event

Also, it looks like we are filling the TITLE with the 'message body' of the text and the DESCRIPTION with the text 'subject'.  This seems backwards though most texts will not include a true subject.  Could we check for a message 'subject' and if there is one fill the title with the subject and description with message body?  If there's no subject, then fill the title with the message body instead?

In general I think this will be a great feature addition to the Mms app...

Patch-set: 5