Update patch set 6
Patch Set 6: (1 inline comment)

See inline comment for extraneous brace.

I see the changes to have message subject fill the title and message body fill the description.  Note that if there is no subject (most cases) the title will be empty.  Are we Ok with that or should we consider filling title with message body IF there is no message subject?

Final thought, are we using the most current calendar API calls? (I honestly don't know but have read references to a new API that went into effect for ICS at 4.0)

Patch-set: 6