Update patch set 6
Patch Set 6:

Re: maybe like this?

The check could be done by the caller or in the function, just have to decide where it make the most sense.  I had considered putting it in the function to put less demand on the caller.  If the Title submitted is empty then fill the title automatically with the first 15 or 20 characters of the description.  (short titles fit better in calendar)  That way the functionality is there for all users of the function. That's only my opinion...

Patch-set: 6