Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

This causes significant lag in the message view when scrolling especially the first time you scroll while viewing a message. It occurs regardless of whether group messaging is toggled on. It does not occur without this set of changes.

Patch-set: 5