Update patch set 7
Patch Set 7: No score

Appears to be working still, scrolling is smooth too

I wonder if this needs to be converted to MMS though. There must have been a reason for Vivek to do that. I'll have to test this without converting to MMS, but I don't know that I'll be able to tonight.

Patch-set: 7