Update patch set 7
Patch Set 7:

Okay here is some feedback.  From what I saw, this feature worked fine on my Galaxy Note with no slowdown, but a few hiccups.  I did put it on as a fresh build, so those that dirty flash, that could be why there may be some slowdown.

This feature works great for iPhones that use group mms, but it does NOT handle blackberry mms at all.  I had a conversation with three people, two had iphones, one had a blackberry.  The blackberry messages kept going to a separate thread.  So something needs to be done about this.

Another thing, if you receive a group text, you cannot use quick reply to reply,  you're only given the option to call.

A minor nitpick is if you're in the sms app, and you get a group text, it will start a new thread, then delete itself and then make a group thread, this is kind of unsightly, so when this feature becomes polished, that would be something to look into.

All in all, it seems solid to me, I did compile this build this morning, so the patches that were applied today were not tested.

Im not a dev, so I can't really help on that front, but I've been following how to get group sms on android for a while, so I hope this comment isn't out of line.

Patch-set: 7