Update patch set 7
Patch Set 7: Looks good to me, but someone else must approve

@Ryan Thanks for the feedback

I'm not sure there is anything we can do about the Blackberry issue. This is the same behavior I witnessed while testing with an android device on the other end that did not have these patches. Some anecdotal evidence from other people suggests that this will work with older Blackberrys but something changed on some newer ones? In their testing this worked fine with a 2 year old Blackberry but a "newer" one would do what you are saying and put it in its own thread.

Quick reply won't work because in order to send a group text, it converts it to an Mms and quick reply does not support Mms.

I have a suspicion that the reason it shows in a thread by itself to begin with is also related to it being an Mms. At the time the phone get's word of an incoming Mms, it must download it first. It wouldn't know the other recipients until its been downloaded. Hence the lag between moving it into a group Mms.

Now, I'm not saying there can't be anything done about any of these, but I think they are over the realm of at least getting initial support for this.

Patch-set: 7