Create change
Patch-set: 1
Change-id: I74e040ba25d8d0dc3fa9c3f6405edc0141482239
Subject:  Mms: Fix not on the screen in some cases.  1. Fix not on the screen, if user not set any lock way(None in the lock setting) and the screen is off.  2. Fix not on the screen, If the pop up dialog show, the screen is off again and receive new message.
Branch: refs/heads/jellybean
Commit: 3d5c8bc2681fb32a5b2c9d5878ffd15a5630f65f