Mms: allows automatic download of mms contents when data connection is off
When a user receives an mms intent, if the data connection is off, the phone
cannot switch the data connection on to download the mms contents. This causes
several problems to users (phone does not deep sleep - drains about 40% battery
in 4/5 hours, users do not get any notifications, etc.).

With this patch, if the data connection is off, the data connection will turn
automatically on and downloads/uploads the mms. Then, connection will turn off.
If data is on, behavior is the same as usual, and data keeps on..

I've tested quite exaustively, and I can send/receive mms as mentioned..

This seems to affect users since android 2.3.7, CM users and vanilla android
users. Can be backported to cm9 and probably cm7.

Change-Id: I1f23f440b786acca492eeefa4c57af543f7276cd