Update patch set 3
Patch Set 3:

Sorry, I was fixing formatting in my Eclipse, so patchsets 1 to 3..

This behavior was the standard behavior up to 2.3.6 and got broken after that. While searching to fix this "error", I read a lot of complains by Android users who want this behavior back. But something must be done, because lots of users (vanilla Android users also) get battery drain because of problems in the Mms app.

But I understand the roaming problem. There is a setting "Data Roaming" which when selected allows data connections while in Roaming. Having that option NOT selected is enough to prevent the data connection? (I am not aware of how that options is enforced).

Patch-set: 3