Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Ok, I checked the CM7 patch [1] and it doesn't seem to be needed here. In CM7, ConnectivityService checked whether data was enabled in startUsingNetworkFeature() and failed if that wasn't the case. This isn't the case for CM10 anymore, thus ConnectivityService should connect to the MMS APN correctly. If a wake lock is held, this seems to be the case for you, as the wake lock is only acquired if startUsingNetworkFeature() returns a success code.

In other words: The data connection for MMS should already be established even _without_ your patch, so the actual problem is probably elsewhere.

Questions:
- What does startUsingNetworkFeature() return for you?
- If it's APN_REQUEST_STARTED: Does the ConnectivityBroadcastReceiver receive a connectivity change afterwards?

When considering the above, I think this is not the correct fix and we should look elsewhere.

[1]https://github.com/maniac103/android_frameworks_base/commit/7b191327001da8e3b5f1bca0346b1aeb27f88a5c

Patch-set: 3