Update patch set 3
Patch Set 3:

Hi Danny,

this is not exactly a fix, it is more a hack. I'll explain.

startUsingNetworkFeature() returns APN_REQUEST_STARTED as you mentioned. Going after the code inside the frameworks/core and frameworks/base, it goes to ConnectivityService.java where it calls a reconnect() on MobileDataStateTracker.java which eventually calls a setEnableApn() which calls a enableApnType() in PhoneProxy.java or PhoneInterfaceManager.

From here I could not identify the path, since my Log.v()'s do not appear in the logcat, although an APN_REQUEST_STARTED is returned from this enableApnType() in MobileDataStateTracker.java. But I do not see an Intent in the possible "alternatives" called by this enableApnType().

This just to say that eventually, I saw in the logcat an info about the getEnabledData() returning false, and that is why I added setEnabledData() although I don't understand exactly what it does (really sets data up or just informs that data connections are allowed - the documentation is not clear about that!).

I agree not to submit this until it can be tested more exhaustively or an alternative is found, although as it is is working. I'll try to pursue this issue later in the day, and report back any findings..

Patch-set: 3