Update patch set 12
Patch Set 12: Looks good to me, but someone else must approve

What do we think here guys? I've had no complaints from anyone testing (quite a few actually) of anything bad that happens because of this.

The one thing that kind of sucks is that it switches to an MMS when you send to multiple people in order for this to work. But that only happens if this is enabled (which is not the default)

Patch-set: 12