Update patch set 1
Patch Set 1:

I only started using Emoji recently and have noticed that the Android look emoji's do tend to look a bit out of place.  Since they can be added via smileys, I have been wondering if we should not have kept the emoji set unchanged.  

IMHO, it should not be a setting - either we revert back to the original emoji set or we keep it as-is.

Patch-set: 1