Update patch set 12
Patch Set 12:

Justo to address the question in a quick manner, other phone platforms (namely iPhones and Blackberries) allow for threaded group messages.    The way it operates currently, if you want to send a text to multiple people, you can put multiple names in the send field, but they get sent as regular SMS to each person.  What this does is convert those SMS to MMS, which contains all the names in the send field, creating a single thread with every contact.  Think of it as a difference between an instant message and a chat room. 

So in android current implementation, if I send a new text and add person A and B to the send field, and A replies, only I see it, and it goes into his connect thread in the text app.  Under threaded MMS, if I send A and B a text, it appears under one "group text" conversation on everyones phone that has group mms enabled, and if A replies, B will get the response as well, and vice versa.

Hope that makes sense.  I personally think this is really a weak point in android, and the fact that it hasn't been addressed by google is really maddening.

Patch-set: 12