Update patch set 12
Patch Set 12:

To add to what I just commented, if an iPhone friend or family tries to send you a group text right now, every persons reply goes to their respective conversation thread.  This becomes really annoying in groups of 3 or more, because every response gets sent to you, and added to their own thread, which makes reading the conversation pretty much impossible when you are trying to find the chronology of the conversation.

This is a long overdue change IMHO, and the fact that it has been sitting on the Commit for a month is unfortunate.

Patch-set: 12