Update patch set 12
Patch Set 12:

Ricardo, for this to work it needs to be converted to an MMS so that it sends the other recipients with the message. This allows the person on the other end to reply back to the group instead of just the person who sent it. SMS isn't able to do this. Mainly the reason that this is an option. If disabled it won't convert it to an MMS.

Patch-set: 12