Update patch set 12
Patch Set 12:

David, obviously make sure its enabled :)

You'll need at least two devices (yours and another) that have this enabled (an iphone5 would work for testing as well)

Send a message to multiple people (it will convert to MMS). When the other person who has group MMS enabled receives it, they'll receive it in a group thread instead of a thread with just your name on it.

When they respond, it will show up in your group thread and everyone in the group will receive the message.

Its pretty much impossible to test with a device that doesn't have this enabled since that device will still receive it in a single thread and wont respond to the entire group because of that.

Patch-set: 12