Update patch set 1
Patch Set 1:

Ricardo why the revert? 

This function was the cherry on the top of the cake for CM! 

Receiving/send mms without enabling mobile data is something that all users are asking for. Most of us don't have a data plan, and programs like Apn Portugal doesn't work on Ics/jb. We have to move it to system via titanium backup or something and every time we install a nightly, my case everyday, we have to repeat the whole process, ie, restore the app, move it to system, add the widget... 

I really don't get the meaning of the revert! 

Regards

Patch-set: 1