Update patch set 1
Patch Set 1:

@Miguel, I merged this by mistake and we had to revert it. The side effect of running MMS as a system user, besides the fact that it is a bad thing to do, is that the MMS sharedpreferences are reset, forcing users to redo their home screen links and recreating the configuration.

We need to go back to the drawing board with this one and see if Danny's suggestion above (quoted below) can be implemented.

"Create a new shared user id type for mms and use it to protect the property. Requires adding it to the UID list in system/core/include/private/android_filesystem_config.h, and to the UID list in frameworks/base/services/java/com/android/server/pm/PackageManagerService.java (in the constructor). This is probably the most invasive (although not overly invasive) alternative, but probably also the cleanest one."

Patch-set: 1