Update patch set 2
Patch Set 2:

See my comment on the other part of the feature.  

Regarding the resetting of the MMS settings, we will probably coordinate the release of this feature with the 4.2 based version of CM10.  Given all the other changes that merging 4.2 will bring, end users will likely be more acceptable of a change requiring a reconfig.

Patch-set: 2