Update patch set 2
Patch Set 2: No score

With Settings.System, we could set the properties in the Mms app, but could not retrieve them in the framework part (check part 2/3 of the patch).

Settings.System is dependent on ContentResolvers (so, is dependent on Context) and in the framework we can not access application Contexts (i think).

Patch-set: 2