Update patch set 1
Patch Set 1:

Auto Enabling is needed for both cases: uploading and downloading. So, If user disables auto-retrieval e does not have also auto-upload. In this way, "auto retrieval" is misleading as name, and should something such as "auto mms transactions"..

But in Gingerbread, if you turn off "auto-retrieval" you will see the mms headers in the mms app and a button for "download". But you won't get a notification. I had a patch for ics which handled that case..

Patch-set: 1