Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Yes, to the extent that the image/photo you are sending via MMS contains exif metadata describing its rotation (the case with photos captured with newer devices, certainly on d2spr).

In addition to making sure the proper orientation is preserved in the image that is actually sent, this patch also makes sure all thumbnails are properly oriented, both the ones appearing in the compose_message window, as well as the ones appearing in the threaded conversation (previously sent).

Patch-set: 1