Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I like it. It looks cleaner.  I was originally going to do it without text but since we only had two buttons originally, the text fit.  Now, with 3 buttons I agree that it needs to be changed.  


Icons need to be properly aligned though.

Patch-set: 1