Update patch set 5
Patch Set 5:

Why was this done at the level of messaging instead of system wide? 

Wouldn't it make more sense to disable scrolling of all notifications instead of restricting it to the messaging app like this?

I feel that a more general system-wide version should be written andt his reverted.

Patch-set: 5