Update patch set 4
Patch Set 4:

I used this for a week and have now reverted back to the version with the text labels.  IMHO it just seems to look better with the labels compared to without.

The guidelines clearly describe the design pattern for Notifications here -> http://developer.android.com/design/patterns/notifications.html

Maybe there is another way we can do this?

Patch-set: 4