Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

The only way I see to avoid ellipsizing of the action text is to verify on every language and on every supported device. That is overdoing it.

As stated before, removing the text altogether goes against Android's Notification Design guidelines. 

Furthermore, "The current icons adequately provide meaning to the actions they reresent." statement appears to be untrue.

From this feedback, it seems better to treat ellipsizing the text as minor nuisance, not something that must be fixed at all costs. Leave the notifications as they are now.

Obs: I'd like to highlight that changing the icon for "mark as read" looks like a good suggestion. But I don't feel confortable copying the icon from a Google proprietary app.

Patch-set: 4