Update patch set 3
Patch Set 3:

There are deletions in strings.xml, I don't know if you intentionally made them.

About the new feature, you re-organized (refactoring, and move code to method) code like you thought it was the best. I'm okay with that, it's a matter of choice. The process is not changed. (and sorry for the part commented, I had forgotten to remove it). 

You made other changes, refactoring, and some optimizations. Should be ok.

Patch-set: 3