Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

We should also rename the 'done' icon which is uesed for the 'mark as read' button in quick reply. It makes no sense that the name is 'done'... We should rename it to 'read' or 'mark_as_read' or something like that.

Patch-set: 3